warning fix

+++ utils/Log/glog.c
@@ -30,7 +30,7 @@ static void __g_print_tag(const char* tag);
 void __g_print_tag(const char* tag)
 {
     const int GLOG_TAG_MAX_SIZE = 5;
-    size_t offset = strlen(tag) > GLOG_TAG_MAX_SIZE + 1 ? 1 : GLOG_TAG_MAX_SIZE - (int)strlen(tag);
+    size_t offset = (uint16_t)(strlen(tag) > GLOG_TAG_MAX_SIZE + 1 ? 1 : GLOG_TAG_MAX_SIZE - (int)strlen(tag));
     gprint("%0" __GLOG_TOSTRING_(__GLOG_MILLIS_SYMBOLS_CNT_) __G_TIME_PRINT_FORMAT "->", getMillis() % __GLOG_MILLIS_TIME_DIV_);
     gprint("%s:%*s", tag, offset, "");
 }
